python/requirements-colab.txt
requests==2.31.0
faiss-cpu==1.8.0.post1
